Rename toxenv files to detect
