+++ docs/detect.rst
