+++ src/tox_travis/detect.py
