+++ tests/test_detect.py
